+++ chef-repo/site-cookbooks/playground-web/recipes/default.rb
@@ -1,11 +1,15 @@
 #
-# Cookbook Name:: playground-application
+# Cookbook Name:: playground-web
 # Recipe:: default
 #
 # Copyright 2017, YOUR_COMPANY_NAME
 #
 # All rights reserved - Do Not Redistribute
 #
+
+package 'php' do
+    action :remove
+end
 
 package 'libicu-devel' do
   action :install